backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.13",
    "beautifulsoup4>=4.13.4",
    "fastapi>=0.115.14",
    "httpx>=0.28.1",
    "httpx-oauth>=0.16.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.68",
    "langchain-tavily>=0.2.7",
    "langchain-upstage>=0.6.0",
    "langgraph>=0.5.1",
    "langgraph-supervisor>=0.0.27",
    "motor>=3.7.1",
    "pyjwt>=2.10.1",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "pytz>=2025.2",
    "requests>=2.32.4",
    "selenium>=4.34.1",
    "uvicorn>=0.35.0",
    "uvloop>=0.21.0",
]

[dependency-groups]
dev = [
    "yapf>=0.43.0",
]
